Refactor code base and split into more module files.
